add data file for repro
likes data isn't particularly sensitive, but it was collected off the atproto firehose without any filtering and any deleted records since capture will still be present here -- i don't want to publish the raw contents directly.

the raw data was processed with a very simple anonymizer pass: swap every real DID with a randomly generated DID, both in the at-URIs and in the record identifiers. i think this is sufficient here, but note that all rkeys are left as-is, and they have a lot of entropy. i don't know how you could usefully use this, but noting it anyway.